Software/Proyecto/requirements.txt
tkinter
mysql-connector-python
fpdf
pandas

#pip install -r requirements.txt
#pip list
